AB#1447 Use a single polling hook that updates the AppContext
