+++ client/src/components/TopBar.tsx
@@ -1,6 +1,4 @@
-import { gql } from "@apollo/client"
 import { resetPages } from "actions"
-import API from "api"
 import AppContext from "components/AppContext"
 import GeneralBanner, {
   GENERAL_BANNER_LEVEL,
@@ -12,19 +10,8 @@ import GeneralBanner, {
 import Header from "components/Header"
 import NoPositionBanner from "components/NoPositionBanner"
 import SecurityBanner from "components/SecurityBanner"
-import React, { useContext, useEffect, useMemo, useRef, useState } from "react"
+import React, { useContext, useEffect, useRef, useState } from "react"
 import { connect } from "react-redux"
-
-const POLLING_INTERVAL = 60_000 // milliseconds
-
-const GQL_GET_ADMIN_SETTINGS = gql`
-  query TopBarAdminSettings {
-    adminSettings {
-      key
-      value
-    }
-  }
-`
 
 interface TopBarProps {
   handleTopbarHeight: (...args: unknown[]) => unknown
@@ -46,47 +33,13 @@ const TopBar = ({
   const [height, setHeight] = useState(0)
   const topbarDiv = useRef()
 
-  const {
-    data: adminData,
-    startPolling,
-    stopPolling
-  } = API.useApiQuery(GQL_GET_ADMIN_SETTINGS)
-
-  useEffect(() => {
-    startPolling?.(POLLING_INTERVAL)
-    return () => stopPolling?.()
-  }, [startPolling, stopPolling])
-
-  const polledSettings = useMemo(() => {
-    const list = adminData?.adminSettings || []
-    if (!list.length) {
-      return undefined
-    }
-    const output = {}
-    for (const setting of list) {
-      output[setting.key] = setting.value
-    }
-    return output
-  }, [adminData])
-
-  const effectiveSettings = polledSettings || appSettings
-  const visibilitySetting = parseInt(
-    effectiveSettings[GENERAL_BANNER_VISIBILITY],
-    10
-  )
-  const bannerOptions = useMemo(
-    () => ({
-      level: effectiveSettings[GENERAL_BANNER_LEVEL],
-      message: effectiveSettings[GENERAL_BANNER_TEXT],
-      title: GENERAL_BANNER_TITLE,
-      visible: bannerVisibility
-    }),
-    [
-      effectiveSettings[GENERAL_BANNER_LEVEL],
-      effectiveSettings[GENERAL_BANNER_TEXT],
-      bannerVisibility
-    ]
-  )
+  const visibilitySetting = parseInt(appSettings[GENERAL_BANNER_VISIBILITY], 10)
+  const bannerOptions = {
+    level: appSettings[GENERAL_BANNER_LEVEL],
+    message: appSettings[GENERAL_BANNER_TEXT],
+    title: GENERAL_BANNER_TITLE,
+    visible: bannerVisibility
+  }
 
   useEffect(() => {
     function updateTopbarHeight() {
